PokeGen/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Pillow>=11.0.0
diffusers>=0.24.0
transformers>=4.35.0
torch>=2.1.0
accelerate>=0.24.0
